Added checks for ActiveRecord::Base.enum support before running related specs

+++ spec/spec_helper.rb
@@ -27,8 +27,10 @@ RSpec.configure do |config|
   end
 
   config.before(:suite) do
-    class UserWithRoleEnum < ActiveRecord::Base; enum role: [:default, :admin] end
-    ActiveRecord::Migration.create_table(:user_with_role_enums) {|t| t.integer :role, default: 0 }
+    if ActiveRecord::Base.respond_to? :enum
+      class UserWithRoleEnum < ActiveRecord::Base; enum role: [:default, :admin] end
+      ActiveRecord::Migration.create_table(:user_with_role_enums) {|t| t.integer :role, default: 0 }
+    end
 
     if defined?(DataMapper)
       # NOTE: eager_loading needs to be on in the app for testing.
@@ -40,7 +42,9 @@ RSpec.configure do |config|
   end
 
   config.after(:suite) do
-    ActiveRecord::Migration.drop_table :user_with_role_enums
+    if ActiveRecord::Base.respond_to? :enum
+      ActiveRecord::Migration.drop_table :user_with_role_enums
+    end
   end
 
   config.before(:each) do
